Move RigidBody from flight_dynamics -> rigid_body
